requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.0
psycopg2-binary==2.9.10
flask-login==0.6.3
Werkzeug>=3.0.3
Jinja2>=3.1.4
python-slugify==8.0.1
Flask-WTF==1.2.2
wtforms==3.2.1
replit>=4.1.0
email_validator==2.2.0
python-dotenv>=1.0.0
flask-limiter==3.8.0
flask-seasurf==2.0.0
flask-talisman==1.1.0
flask-mail==0.10.0
flask-session==0.8.0
sentry-sdk==2.17.0
Flask-Caching==2.3.0
redis==5.2.0
gunicorn>=22.0.0
gevent>=23.9.1
psycogreen>=1.0.1

# Email Service (Resend)
resend>=2.0.0

# Native Statement System (Phase 1 & 3)
# Note: scipy deferred - Wilson score can be calculated without it
# scipy will be added later with proper binary wheels
scikit-learn>=1.3.0
numpy>=1.24.0

# Background Tasks (Phase 3)
APScheduler==3.10.4

# Encryption for API keys (Phase 4)
# Upper bound required: pyseto (replit dependency) requires cryptography<44
cryptography>=43.0.1,<44.0.0
pandas>=2.0.0
dnspython>=2.6.1
markdown>=3.6
bleach>=6.1.0

# Social Media Posting
atproto>=0.0.30
tweepy>=4.14.0

# Briefing System v2 - Ingestion
pypdf>=3.0.0
python-docx>=1.0.0
beautifulsoup4>=4.12.0
readability-lxml>=0.8.1
pytz>=2023.3

# TLS CA bundle (journey link checker, urllib clients on constrained hosts)
certifi>=2024.2.2

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Pin protobuf to compatible version (required by replit, googleapis packages)
protobuf>=4.21.8,<5.0.0
